% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsocsim.R
\name{create_simulation_folder}
\alias{create_simulation_folder}
\title{Create a directory structure for the simulation}
\usage{
create_simulation_folder(basedir = NULL, simdir = NULL)
}
\arguments{
\item{basedir}{A string. Optional. First-level directory where the
simulation-specific directory will be created. Defaults to
\code{file.path(tempdir(), "socsim")}.}

\item{simdir}{A string. Optional. Simulation-specific directory which will
be created within 'basedir'. Defaults to 'socsim_sim_' plus a random
component created with \code{\link[=tempfile]{tempfile()}}.}
}
\value{
A string. The full path to the simulation-specific directory.
}
\description{
Create a two-level directory structure. If the first-level argument is NULL,
we look for and, if needed, create the directory 'socsim' in the current
temporary directory. If the second-level argument is NULL, we create a directory
named like 'socsim_sim_' followed by a random component in the first-level
directory.
}
