% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsocsim.R
\name{create_sup_file}
\alias{create_sup_file}
\title{create a basic .sup file for a simulation
the simulation is only a simple one
the file will be saved into the sim-folder}
\usage{
create_sup_file(simfolder, simname)
}
\arguments{
\item{simfolder}{the folder where the sup-file will be saved}

\item{simname}{the name of the simulation}
}
\value{
sup.fn the filename of the supplementary file
which is needed to start the simulation
}
\description{
create a basic .sup file for a simulation
the simulation is only a simple one
the file will be saved into the sim-folder
}
