% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_rates.R
\name{estimate_mortality_rates}
\alias{estimate_mortality_rates}
\title{Estimate yearly age-specific mortality rates (ASMR) from a SOCSIM-generated
population file}
\usage{
estimate_mortality_rates(
  opop,
  final_sim_year,
  year_min,
  year_max,
  year_group,
  age_max_mort,
  age_group
)
}
\arguments{
\item{opop}{An R object from SOCSIM microsimulation output (population file).}

\item{final_sim_year}{numeric. Final simulated year in 'real world' time (
used to convert 'SOCSIM time' to 'real world' time.)}

\item{year_min}{numeric. Lower-bound year for which rate should be estimated.}

\item{year_max}{numeric. Upper-bound year for which rate should be estimated.}

\item{year_group}{numeric. Size of year groups to estimate rate (year_group=1
will produce single-year estimates)}

\item{age_max_mort}{numeric. Maximum age for estimating mortality.}

\item{age_group}{numeric. Size of age groups to estimate rate (age_group=1
will produce single-age estimates)}
}
\value{
A data frame with columns \code{year}, \code{sex}, \code{age}, and \code{socsim}.
\code{year} is a factor describing the grouped calendar-year interval,
\code{sex} is \code{"male"} or \code{"female"}, \code{age} is a factor describing the age
interval, and \code{socsim} is the estimated mortality rate for that cell.
}
\description{
Given a population file ('opop') generated by rsocsim, the
function estimates (yearly) age-specific mortality rates.
}
\details{
The \code{final_sim_year} can be obtained from the .sup file and must
refer to to a real-world year.

Grouped year and age ranges (i.e., if \code{year_group > 1} or \code{age_group > 1})
are created as [year;year+year_group). For \code{age_group > 1}, mortality rates
are split into an infant group [0,1) and then grouped ages [1, age_group),
[age_group, age_group + age_group), and so on.
}
\examples{
opop <- data.frame(
  pid = 1:6,
  fem = c(1, 0, 1, 0, 1, 1),
  group = 1,
  nev = 0,
  dob = c(120, 120, 336, 348, 180, 360),
  mom = c(0, 0, 1, 1, 0, 5),
  pop = c(0, 0, 2, 2, 0, 2),
  nesibm = 0,
  nesibp = 0,
  lborn = 0,
  marid = 0,
  mstat = 0,
  dod = c(0, 300, 0, 0, 0, 0),
  fmult = 0
)

asmr <- estimate_mortality_rates(opop = opop,
                     final_sim_year = 2021,
                     year_min = 1995,
                     year_max = 2000,
                     year_group = 5,
                     age_max_mort = 100,
                     age_group = 5)
head(asmr)
}
