% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsocsim.R
\name{get_supplement_content}
\alias{get_supplement_content}
\title{read the content of the supplement file}
\usage{
get_supplement_content(simfolder, sup_fn)
}
\arguments{
\item{simfolder}{base folder of the simulation}

\item{simname}{name of the .sup-file}
}
\value{
the content of the supplement file as a string (TODO: Now it 
returns a list of lines instead of a single string)
}
\description{
read the content of the supplement file
}
