% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_kin.R
\name{getKin}
\alias{getKin}
\title{Identify members of a kin network for an individual or individuals of 
interest.}
\usage{
getKin(opop = opop, KidsOf = KidsOf, pid, extra_kintypes, kin_by_sex)
}
\arguments{
\item{opop}{An R object from SOCSIM microsimulation output (population file).}

\item{pid}{A character vector of person IDs, indicating persons of interest
for whom these kin networks should be identified.}

\item{extra_kintypes}{A vector of character values indicating which
additional types of kin should be obtained. For reasons of computational
efficiency, the function will by default only identify an individual's
 great-grandparents ("ggparents" in function output), 
 grandparents ("gparents"), parents, 
 siblings, spouse, children, and grandchildren ("gchildren"). 
 However, by selecting one or more of the following 
 kin types, the kin network generated will also include these individuals:
 \begin{itemize}
 \item "gunclesaunts": Great-uncles and great-aunts
 \item "unclesaunts": Uncles and aunts
 \item "firstcousins": First cousins (Children of uncles and aunts)
 \item "niblings": Nieces and nephews (Children of siblings)
 \item "inlaws": Parents-in-law (parents of spouse) and brothers and sisters
 in law (siblings of spouse and spouse of siblings)
 \end{itemize}}

\item{kin_by_sex}{A logical value indicating whether output should include
kin relations additionally disaggregated by the sex of the relative. Setting
this value to TRUE will result in additional objects being generated
to identify individuals' relatives by sex.}

\item{omar}{An R object from SOCSIM microsimulation output (marriage file).}
}
\value{
An R object containing a list of lists with person IDs
of kin, organized by relationship. These person ID values will be named 
based on the person of interest with whom they are associated.
For example, for a list named "parents", the values will be person IDs of
the parents of individuals of interest. These values will be named according
to their children's IDs (given that their children are, in this case, 
the persons of interest provided to the function input).
With kin_by_sex set to TRUE and extra_kintypes set to c(c("gunclesaunts",
 "unclesaunts", "firstcousins", "niblings", "inlaws")), 
 the full list of kin relations identified are:
 \begin{itemize}
 \item "ggparents": great-grandparents
 \item "ggmothers": great-grandmothers
 \item "ggfathers": great-grandfathers
 \item "gparents": grandparents
 \item "gmothers": grandmothers
 \item "gfathers": grandfathers
 \item "gunclesaunts": great-uncles and great-aunts
 \item "guncles": great-uncles
 \item "gaunts": great-aunts
 \item "parents": parents
 \item "mother": mother
 \item "father": father
 \item "unclesaunts": uncles and aunts (siblings of parents)
 \item "uncles": uncles 
 \item "aunts": aunts
 \item "siblings": siblings
 \item "sisters": sisters
 \item "brothers": brothers
 \item "firstcousins": first cousins
 \item "firstcousinsfemale": female first cousins
 \item "firstcousinsmale": male first cousins
 \item "children": children
 \item "daughters": daughters
 \item "sons": sons
 \item "gchildren": grandchildren
 \item "gdaughters": granddaughters
 \item "gsons": grandsons
 \item "niblings": nephews and nieces
 \item "nieces": nieces
 \item "nephews": nephews
 \item "spouse": spouse (based on final marriage, in the case of 
 multiple marriages)
 \item "parentsinlaw": parents-in-law
 \item "motherinlaw": mother-in-law
 \item "fatherinlaw": father-in-law
 \item "siblingsinlaw": brothers and sisters in law
 \item "sistersinlaw": sisters-in-law
 \item "brothersinlaw": brothers-in-law
 \end{itemize}
}
\description{
Identify members of a kin network for an individual or individuals of 
interest.
}
\examples{
\dontrun{
#Individuals of interest
pid <- c("10111", "10211", "10311")
#Obtain partial kinship network, with omar and opop already in R environment
kin_network <- getKin(opop = opop, omar = omar, pid = pid, 
extra_kintypes = c("unclesaunts", "niblings"), kin_by_sex = TRUE)
}

}
