% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_kin.R
\name{retrieve_kin}
\alias{retrieve_kin}
\title{Identify members of a kin network for an individual or individuals of
interest.}
\usage{
retrieve_kin(
  opop,
  omar,
  pid,
  extra_kintypes = character(),
  kin_by_sex = FALSE,
  KidsOf = NULL
)
}
\arguments{
\item{opop}{An R object from SOCSIM microsimulation output (population file). Create this object with the function read_opop().}

\item{omar}{An R object from SOCSIM microsimulation output (marriage file). Create this object with the function read_omar().}

\item{pid}{A vector of person IDs, indicating persons of interest
for whom these kin networks should be identified.}

\item{extra_kintypes}{A vector of character values indicating which
additional types of kin should be obtained. For reasons of computational
efficiency, the function will by default only identify an individual's
great-grandparents ("ggparents" in function output),
grandparents ("gparents"), parents,
siblings, spouse, children, and grandchildren ("gchildren").
However, by selecting one or more of the following
kin types, the kin network generated will also include these individuals:
\itemize{
\item "gunclesaunts": Great-uncles and great-aunts
\item "unclesaunts": Uncles and aunts
\item "firstcousins": First cousins (Children of uncles and aunts)
\item "niblings": Nieces and nephews (Children of siblings)
\item "inlaws": Parents-in-law (parents of spouse) and brothers and sisters
in law (siblings of spouse and spouse of siblings)
}}

\item{kin_by_sex}{A logical value indicating whether output should include
kin relations additionally disaggregated by the sex of the relative. Setting
this value to TRUE will result in additional objects being generated
to identify individuals' relatives by sex.}

\item{KidsOf}{An optional precomputed list object containing the children of
each person in the population. If \code{NULL}, it is built from \code{opop}.}
}
\value{
A named list whose components are kinship categories such as
\code{parents}, \code{siblings}, or \code{children}. Each component is itself a named list
of integer person IDs, organized by relationship. These person ID values will be named
based on the person of interest with whom they are associated.
For example, for a list named "parents", the values will be person IDs of
the parents of individuals of interest. These values will be named according
to their children's IDs (given that their children are, in this case,
the persons of interest provided to the function input).
With kin_by_sex set to TRUE and extra_kintypes set to c(c("gunclesaunts",
"unclesaunts", "firstcousins", "niblings", "inlaws")),
the full list of kin relations identified are:
\itemize{
\item "ggparents": great-grandparents
\item "ggmothers": great-grandmothers
\item "ggfathers": great-grandfathers
\item "gparents": grandparents
\item "gmothers": grandmothers
\item "gfathers": grandfathers
\item "gunclesaunts": great-uncles and great-aunts
\item "guncles": great-uncles
\item "gaunts": great-aunts
\item "parents": parents
\item "mother": mother
\item "father": father
\item "unclesaunts": uncles and aunts (siblings of parents)
\item "uncles": uncles
\item "aunts": aunts
\item "siblings": siblings
\item "sisters": sisters
\item "brothers": brothers
\item "firstcousins": first cousins
\item "firstcousinsfemale": female first cousins
\item "firstcousinsmale": male first cousins
\item "children": children
\item "daughters": daughters
\item "sons": sons
\item "gchildren": grandchildren
\item "gdaughters": granddaughters
\item "gsons": grandsons
\item "niblings": nephews and nieces
\item "nieces": nieces
\item "nephews": nephews
\item "spouse": spouse (based on final marriage, in the case of
multiple marriages)
\item "parentsinlaw": parents-in-law
\item "motherinlaw": mother-in-law
\item "fatherinlaw": father-in-law
\item "siblingsinlaw": brothers and sisters in law
\item "sistersinlaw": sisters-in-law
\item "brothersinlaw": brothers-in-law
}
}
\description{
Identify members of a kin network for an individual or individuals of
interest.
}
\examples{
opop <- data.frame(
  pid = 1:4,
  fem = c(1, 0, 1, 0),
  group = 1,
  nev = 0,
  dob = c(120, 120, 300, 300),
  mom = c(0, 0, 1, 1),
  pop = c(0, 0, 2, 2),
  nesibm = 0,
  nesibp = 0,
  lborn = 0,
  marid = c(1, 1, 0, 0),
  mstat = c(4, 4, 1, 1),
  dod = 0,
  fmult = 0
)
omar <- data.frame(
  mid = 1,
  wpid = 1,
  hpid = 2,
  dstart = 0,
  dend = 0,
  rend = 0,
  wprior = 0,
  hprior = 0
)

kin_network <- retrieve_kin(
  opop = opop,
  omar = omar,
  pid = 3,
  extra_kintypes = c("niblings", "inlaws"),
  kin_by_sex = TRUE
)
kin_network$parents[[1]]

}
