\name{rsoc-package}
\alias{rsoc-package}
\alias{rsoc}
\docType{package}
\title{
  SOCSIM Package for R. Run Social microsimulations directly from R.
}
\description{
  This package makes it easy to install socsim, run simulations, and
  read the results.
}
\details{
  
}
\author{
Maintainer: Tom Theile <theile@mpidr.de>  and Diego  Alburez-Gutierrez <alburezgutierrez@demogr.mpg.de>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{ 
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
