% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsocsim.R
\name{run1simulationwithfile_from_binary}
\alias{run1simulationwithfile_from_binary}
\title{Run a single socsim-simulation with a socsim binary}
\usage{
run1simulationwithfile_from_binary(
  folder,
  supfile,
  seed = "42",
  compatibility_mode = "1",
  socsim_path = NULL
)
}
\arguments{
\item{folder}{base-directory of the simulation.}

\item{supfile}{the .sup file to start the simulation, relative to the
folder}

\item{seed}{RNG seed as string, Default="42"}

\item{compatibility_mode}{compatibility mode as string, Default="1"}

\item{socsim_path}{path+filename of a socsim-executable. Download one from https://github.com/tomthe/socsim/releases/download/0.3/socsim.exe}
}
\value{
The results will be written into the specified folder
}
\description{
Run a single socsim-simulation with a socsim binary
}
