% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsocsim.R
\name{simulation_time_to_years}
\alias{simulation_time_to_years}
\title{Calculate for how many years the simulation ran}
\usage{
simulation_time_to_years(simulation_time, pre_simulation_time, start_year)
}
\arguments{
\item{simulation_time}{An integer. The number of periods (months) the
simulation ran.}

\item{pre_simulation_time}{An integer. The number of periods (months) the
simulation ran before getting to a stable population. This is subtracted
from 'simulation_time' in order to arrive at the "real" simulation time}

\item{start_year}{An integer. The year the simulation started.}
}
\value{
A numeric scalar giving the calendar year reached at the end of the
simulated period after subtracting \code{pre_simulation_time / 12}. The value
can include a fractional year.
}
\description{
Calculate for how many years the simulation ran
}
