% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsocsim.R
\name{create_simulation_folder}
\alias{create_simulation_folder}
\title{create a folder in the user-dir of the current user in the socsim-subfolder}
\usage{
create_simulation_folder(simulation_name = NULL, basefolder = NULL)
}
\arguments{
\item{simulation_name}{optional name for the simulation}

\item{basefolder}{optional base directory where the folder will be created}
}
\value{
the path to the folder
}
\description{
create a folder in the user-dir of the current user in the socsim-subfolder
}
