% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsocsim.R
\name{create_sup_file}
\alias{create_sup_file}
\title{Create a basic .sup file for a simulation}
\usage{
create_sup_file(simdir, simname)
}
\arguments{
\item{simdir}{A string. The directory where the .sup file will be saved.}

\item{simname}{A string. The name of the simulation.}
}
\value{
A string. The filename of the supervisory file which is needed to
start the simulation.
}
\description{
The simulation is only a simple one. The file will be saved into the
directory 'simdir'.
}
