% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_omar}
\alias{read_omar}
\title{Title}
\usage{
read_omar(
  folder = NULL,
  supfile = "socsim.sup",
  seed = 42,
  suffix = "",
  fn = NULL
)
}
\arguments{
\item{folder}{simulation base folder ("~/socsim/simulation_235/")}

\item{supfile}{name of supplement-file ("socsim.sup")}

\item{seed}{random number seed (42)}

\item{suffix}{optional suffix for the results-directory (default="")}

\item{fn}{complete path to the file. If not provided, it will be created from the other arguments}
}
\value{
Data frame with the information of the output population file
}
\description{
| **position** | **name** | **description**                                                    |
| -----------: | :------- | :----------------------------------------------------------------- |
|            1 | mid      | Marriage id number (unique sequential integer)                     |
|            2 | wpid     | Wife’s person id                                                   |
|            3 | hpid     | Husband’s person id                                                |
|            4 | dstart   | Date marriage began                                                |
|            5 | dend     | Date marriage ended or zero if still in force at end of simulation |
|            6 | rend     | Reason marriage ended 2 = divorce; 3 = death of one partner        |
|            7 | wprior   | Marriage id of wife’s next most recent prior marriage              |
|            8 | hprior   | Marriage id of husband’s next most recent prior marriage           |
}
\details{
you can either provide the complete `path` to the file or the `folder`, supfilename, seed and suffix with which you
started the simulation
}
