% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_omar}
\alias{read_omar}
\title{Read output marriage file into a data frame}
\usage{
read_omar(
  folder = NULL,
  supfile = "socsim.sup",
  seed = 42,
  suffix = "",
  fn = NULL
)
}
\arguments{
\item{folder}{simulation base folder ("~/socsim/simulation_235/")}

\item{supfile}{name of supplement-file ("socsim.sup")}

\item{seed}{random number seed (42)}

\item{suffix}{optional suffix for the results-directory (default="")}

\item{fn}{complete path to the file. If not provided, it will be created from the other arguments}
}
\value{
Data frame with the information of the output population file
}
\description{
\tabular{rll}{
\link{,1} \tab 1   \tab mid \tab Marriage id number (unique sequential integer) \cr
\link{,2} \tab 2   \tab wpid \tab Wife’s person id \cr
\link{,3} \tab 3   \tab hpid \tab Husband’s person id \cr
\link{,4} \tab 4   \tab dstart \tab Date marriage began \cr
\link{,5} \tab 5   \tab dend \tab Date marriage ended or zero if still in force at end of simulation \cr
\link{,6} \tab 6   \tab rend \tab Reason marriage ended 2 = divorce; 3 = death of one partner \cr
\link{,7} \tab 7   \tab wprior \tab Marriage id of wife’s next most recent prior marriage \cr
\link{,8} \tab 8   \tab hprior \tab Marriage id of husband’s next most recent prior marriage \cr
}
}
\details{
you can either provide the complete \code{path} to the file or the \code{folder}, supfilename, seed and suffix with which you
started the simulation
}
