% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_opop}
\alias{read_opop}
\title{Read output population file into a data frame}
\usage{
read_opop(
  folder = NULL,
  supfile = "socsim.sup",
  seed = 42,
  suffix = "",
  fn = NULL
)
}
\arguments{
\item{folder}{simulation base folder ("~/socsim/simulation_235/")}

\item{supfile}{name of supplement-file ("socsim.sup")}

\item{seed}{random number seed (42)}

\item{suffix}{optional suffix for the results-directory (default="")}

\item{fn}{complete path to the file. If not provided, it will be created from the other arguments}
}
\value{
Data frame with the information of the output population file
}
\description{
after the end of the simulation, socsim writes every person of the simulation into
a file called result.opop. This file contains the following information about
each person:
}
\details{
| **position** | **name** | **description**                                                                       |
| -----------: | :------- | :------------------------------------------------------------------------------------ |
|            1 | pid      | Person id unique identifier assigned as integer in birth order                        |
|            2 | fem      | 1 if female 0 if male                                                                 |
|            3 | group    | Group identifier 1..60 current group membership of individual                         |
|            4 | nev      | Next scheduled event                                                                  |
|            5 | dob      | Date of birth integer month number                                                    |
|            6 | mom      | Person id of mother                                                                   |
|            7 | pop      | Person id of father                                                                   |
|            8 | nesibm   | Person id of next eldest sibling through mother                                       |
|            9 | nesibp   | Person id of next eldest sibling through father                                       |
|           10 | lborn    | Person id of last born child                                                          |
|           11 | marid    | Id of marriage in .omar file                                                          |
|           12 | mstat    | Marital status at end of simulation integer 1=single;2=divorced; 3=widowed; 4=married |
|           13 | dod      | Date of death or 0 if alive at end of simulation                                      |
|           14 | fmult    | Fertility multiplier                                                                  |

you can either provide the complete `path` to the file or the `folder`, supfilename, seed and suffix with which you
started the simulation
}
