% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_results.R
\name{read_opop}
\alias{read_opop}
\title{Read output population file into a data frame}
\usage{
read_opop(
  folder = NULL,
  supfile = "socsim.sup",
  seed = 42,
  suffix = "",
  fn = NULL
)
}
\arguments{
\item{folder}{simulation base folder ("~/socsim/simulation_235/")}

\item{supfile}{name of supervisory-file ("socsim.sup")}

\item{seed}{random number seed (42)}

\item{suffix}{optional suffix for the results-directory (default="")}

\item{fn}{complete path to the file. If not provided, it will be created from the other arguments}
}
\value{
Data frame with the information of the output population file
}
\description{
after the end of the simulation, socsim writes every person of the simulation into
a file called result.opop                                                              |
}
\details{
\tabular{rll}{
1   \tab pid \tab Person id unique identifier assigned as integer in birth order \cr
2   \tab fem \tab 1 if female 0 if male \cr
3   \tab group \tab Group identifier 1..60 current group membership of individual \cr
4   \tab nev \tab Next scheduled event \cr
5   \tab dob \tab Date of birth integer month number \cr
6   \tab mom \tab Person id of mother \cr
7   \tab pop \tab Person id of father \cr
8   \tab nesibm \tab Person id of next eldest sibling through mother \cr
9   \tab nesibp \tab Person id of next eldest sibling through father \cr
10  \tab lborn \tab Person id of last born child \cr
11  \tab marid \tab Id of marriage in .omar file \cr
12  \tab mstat \tab Marital status at end of simulation integer 1=single;2=divorced; 3=widowed; 4=married \cr
13  \tab dod \tab Date of death or 0 if alive at end of simulation \cr
14  \tab fmult \tab Fertility multiplier \cr
}

This table explains the columns of the opop file and the columns of the output data frame.
You can either provide the complete \code{path} to the file or the \code{folder}, supfilename, seed and suffix with which you
started the simulation
}
