% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsocsim.R
\name{run1simulationwithfile}
\alias{run1simulationwithfile}
\title{Run a socsim-simulation in a seperate process}
\usage{
run1simulationwithfile(rootfolder, folder, supfile, seed)
}
\arguments{
\item{rootfolder}{rootfolder...}

\item{folder}{folder name of the simulation}

\item{supfile}{the .sup file to start the simulation}

\item{seed}{RNG seed}
}
\value{
basically nothing
}
\description{
Run a socsim-simulation in a seperate process
}
