% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsocsim.R
\name{run1simulationwithfile_inprocess}
\alias{run1simulationwithfile_inprocess}
\title{Run a socsim-simulation in the r-process}
\usage{
run1simulationwithfile_inprocess(rootfolder, supfile, seed)
}
\arguments{
\item{rootfolder}{rootfolder  name of the simulation}

\item{supfile}{the .sup file to start the simulation}

\item{seed}{RNG seed}
}
\value{
The results will be written into the specified folder
}
\description{
Run a socsim-simulation in the r-process
}
