% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startsocsim.R
\name{simulation_time_to_years}
\alias{simulation_time_to_years}
\title{simulation_time_to_years
convert the time measures.}
\usage{
simulation_time_to_years(simulation_time, pre_simulation_time, start_year)
}
\arguments{
\item{simulation}{time}

\item{pre-simulation-time}{- how long the simulation ran to get a stable population}

\item{start-year}{- the year the simulation started}
}
\value{
year, a number like 2022.2
}
\description{
simulation_time_to_years
convert the time measures.
}
